Added initial DataStorage support for more stats Added agrade_levels stat Added pacifist_levels stat Added dlc_chaliced_levels stat Added PlayerLevelDataIndex for performant querying of level data LevelBits are unique Added more LevelMap id's for some non-check levels

+++ CupheadArchipelago/AP/APSlotData.cs
@@ -52,7 +52,7 @@ namespace CupheadArchipelago.AP {
             version = GetAPSlotDataValue<long>(slotData, "version");
             world_version = GetAPSlotDataVersion(slotData, "world_version");
             feature_bit_reqs = GetOptionalAPSlotDataValue<long>(slotData, "feature_bit_reqs", 0);
-            level_map = new LevelMap(GetAPSlotDataDeserializedValue<Dictionary<long, long>>(slotData, "level_map"));
+            level_map = new LevelMap(GetAPSlotDataDeserializedValue<Dictionary<int, int>>(slotData, "level_map"));
             shop_map = GetAPShopMap(slotData);
             //Logging.Log($"shop_map: {shop_map}");
             use_dlc = GetAPSlotDataValue<bool>(slotData, "use_dlc");
